:tada: Add word boundary navigation
