--- frontend/src/app/render_wasm/api/shared.js
@@ -1,256 +1,0 @@
-export const GrowType = {
-  "fixed": 0,
-  "auto-width": 1,
-  "auto-height": 2,
-};
-
-export const RawBlendMode = {
-  "normal": 3,
-  "screen": 14,
-  "overlay": 15,
-  "darken": 16,
-  "lighten": 17,
-  "color-dodge": 18,
-  "color-burn": 19,
-  "hard-light": 20,
-  "soft-light": 21,
-  "difference": 22,
-  "exclusion": 23,
-  "multiply": 24,
-  "hue": 25,
-  "saturation": 26,
-  "color": 27,
-  "luminosity": 28,
-};
-
-export const RawBlurType = {
-  "layer-blur": 0,
-};
-
-export const RawFillData = {
-  "solid": 0,
-  "linear": 1,
-  "radial": 2,
-  "image": 3,
-};
-
-export const RawFontStyle = {
-  "normal": 0,
-  "italic": 1,
-};
-
-export const RawAlignItems = {
-  "start": 0,
-  "end": 1,
-  "center": 2,
-  "stretch": 3,
-};
-
-export const RawAlignContent = {
-  "start": 0,
-  "end": 1,
-  "center": 2,
-  "space-between": 3,
-  "space-around": 4,
-  "space-evenly": 5,
-  "stretch": 6,
-};
-
-export const RawJustifyItems = {
-  "start": 0,
-  "end": 1,
-  "center": 2,
-  "stretch": 3,
-};
-
-export const RawJustifyContent = {
-  "start": 0,
-  "end": 1,
-  "center": 2,
-  "space-between": 3,
-  "space-around": 4,
-  "space-evenly": 5,
-  "stretch": 6,
-};
-
-export const RawJustifySelf = {
-  "none": 0,
-  "auto": 1,
-  "start": 2,
-  "end": 3,
-  "center": 4,
-  "stretch": 5,
-};
-
-export const RawAlignSelf = {
-  "none": 0,
-  "auto": 1,
-  "start": 2,
-  "end": 3,
-  "center": 4,
-  "stretch": 5,
-};
-
-export const RawVerticalAlign = {
-  "top": 0,
-  "center": 1,
-  "bottom": 2,
-};
-
-export const RawConstraintH = {
-  "left": 0,
-  "right": 1,
-  "leftright": 2,
-  "center": 3,
-  "scale": 4,
-};
-
-export const RawConstraintV = {
-  "top": 0,
-  "bottom": 1,
-  "topbottom": 2,
-  "center": 3,
-  "scale": 4,
-};
-
-export const RawFlexDirection = {
-  "row": 0,
-  "row-reverse": 1,
-  "column": 2,
-  "column-reverse": 3,
-};
-
-export const RawWrapType = {
-  "wrap": 0,
-  "nowrap": 1,
-};
-
-export const RawGridDirection = {
-  "row": 0,
-  "column": 1,
-};
-
-export const RawGridTrackType = {
-  "percent": 0,
-  "flex": 1,
-  "auto": 2,
-  "fixed": 3,
-};
-
-export const RawSizing = {
-  "fill": 0,
-  "fix": 1,
-  "auto": 2,
-};
-
-export const RawBoolType = {
-  "union": 0,
-  "difference": 1,
-  "intersection": 2,
-  "exclusion": 3,
-};
-
-export const RawSegmentData = {
-  "move-to": 1,
-  "line-to": 2,
-  "curve-to": 3,
-  "close": 4,
-};
-
-export const RawShadowStyle = {
-  "drop-shadow": 0,
-  "inner-shadow": 1,
-};
-
-export const RawShapeType = {
-  "frame": 0,
-  "group": 1,
-  "bool": 2,
-  "rect": 3,
-  "path": 4,
-  "text": 5,
-  "circle": 6,
-  "svg-raw": 7,
-};
-
-export const RawStrokeStyle = {
-  "solid": 0,
-  "dotted": 1,
-  "dashed": 2,
-  "mixed": 3,
-};
-
-export const RawStrokeCap = {
-  "none": 0,
-  "line-arrow": 1,
-  "triangle-arrow": 2,
-  "square-marker": 3,
-  "circle-marker": 4,
-  "diamond-marker": 5,
-  "round": 6,
-  "square": 7,
-};
-
-export const RawFillRule = {
-  "nonzero": 0,
-  "evenodd": 1,
-};
-
-export const RawStrokeLineCap = {
-  "butt": 0,
-  "round": 1,
-  "square": 2,
-};
-
-export const RawStrokeLineJoin = {
-  "miter": 0,
-  "round": 1,
-  "bevel": 2,
-};
-
-export const RawTextAlign = {
-  "left": 0,
-  "center": 1,
-  "right": 2,
-  "justify": 3,
-};
-
-export const RawTextDirection = {
-  "ltr": 0,
-  "rtl": 1,
-};
-
-export const RawTextDecoration = {
-  "none": 0,
-  "underline": 1,
-  "line-through": 2,
-  "overline": 3,
-};
-
-export const RawTextTransform = {
-  "none": 0,
-  "uppercase": 1,
-  "lowercase": 2,
-  "capitalize": 3,
-};
-
-export const RawGrowType = {
-  "fixed": 0,
-  "auto-width": 1,
-  "auto-height": 2,
-};
-
-export const CursorDirection = {
-  "backward": 0,
-  "forward": 1,
-  "line-before": 2,
-  "line-after": 3,
-  "line-start": 4,
-  "line-end": 5,
-};
-
-export const RawTransformEntryKind = {
-  "parent": 0,
-  "child": 1,
-};
-
